+++ hot/hot-server.js
@@ -30,3 +30,23 @@ WebApp.connectHandlers.use(function(req, res, next) {
   res.writeHead(200, {'Content-Type':'text/javascript'});
   res.end(bundle.contents, 'utf8');
 });
+
+Meteor.methods({
+  '__hot.reload': function() {
+    forceReload();
+  }
+});
+
+var projRoot = process.cwd().substr(0,
+  process.cwd().indexOf('/.meteor/local/build'));
+
+var fs = Npm.require('fs');
+var path = Npm.require('path');
+
+// XXX any better way to do this?
+function forceReload() {
+  fs.writeFileSync(path.join(projRoot, 'client', 'hot-force-reload.js'),
+    '// Temporary measure to force a real reload\n' +
+    'var UNIQUE = "' + Random.hexString(16) + '";\n');
+  hot.col.remove({});
+}